+++ benchmark/matmul/matmul_interface.hpp
@@ -1,5 +1,5 @@
 //
-// SPDX-FileCopyrightText: Copyright 2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: Copyright 2025-2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
 //
 // SPDX-License-Identifier: Apache-2.0
 //
@@ -80,6 +80,19 @@ struct MatMulBlockwiseDynamicQuantInterface {
         float* dst,                                    //
         size_t dst_stride_row, size_t dst_stride_col,  //
         float clamp_min, float clamp_max);
+};
+
+/// Abstraction for the Matrix Multiplication micro-kernel with dynamic blockwise quantization and float destination
+/// buffer with look up table in the interface
+struct MatMulBlockwiseDynamicQuantLutInterface {
+    void (*run_matmul)(
+        size_t m, size_t n, size_t k,                  //
+        const void* lhs_packed,                        //
+        const void* rhs_packed,                        //
+        float* dst,                                    //
+        size_t dst_stride_row, size_t dst_stride_col,  //
+        float clamp_min, float clamp_max,              //
+        const int32_t* lut);
 };
 
 }  // namespace kai::benchmark