+++ CHANGELOG.md
@@ -10,6 +10,11 @@ KleidiAI follows the [Semantic Versioning](https://semver.org/) specification fo
 
 ## Upcoming Release
 
+- New SME2 micro-kernels
+  - Matrix Multiplication (1xN) Micro-Kernel of QAI8DXP LHS and QSU2CXP RHS with F32 output.
+  - Matrix Multiplication (MxN) Micro-Kernel of QAI8DXP LHS and QSU2CXP RHS with F32 output.
+- New Packing kernel for QSU2CXP RHS.
+
 ## v1.21.0
 
 - New SVE micro-kernels